+++ apps/waitlist/lib/scoring/getWaitlistSlot.ts
@@ -1,0 +1,10 @@
+import type { ConnectWaitlistSlot } from '@charmverse/core/prisma-client';
+import { prisma } from '@charmverse/core/prisma-client';
+
+export function getWaitlistSlot({ fid }: { fid: number }): Promise<ConnectWaitlistSlot> {
+  return prisma.connectWaitlistSlot.findUniqueOrThrow({
+    where: {
+      fid
+    }
+  });
+}
