+++ apps/waitlist/theme/colors.ts
@@ -1,3 +1,5 @@
+import { lighten } from '@mui/material';
+
 export const brandColor = '#8742FF'; // purple
 
 // dark mode
@@ -8,3 +10,8 @@ export const backgroundColorDarkMode = '#191919';
 export const backgroundLightColorDarkMode = '#262626';
 export const inputBackgroundDarkMode = 'rgba(255, 255, 255, 0.055)';
 export const inputBorderDarkMode = 'rgba(255, 255, 255, 0.055)';
+
+// Farcaster specific
+export const farcasterBrandColor = '#8465CB';
+export const farcasterBrandColorLight = lighten(farcasterBrandColor, 0.9);
+export const farcasterBrandColorDark = '#6944ba';
